X-Site support on Minikube #125
* Add manual test with 2 Minikube profiles to verify that x-site works.
* It verifies that x-site view gets formed.
* Verify that a cache put is reflected in remote site.
* Use Service.ExternalIPs for configuring Minikube's external IP
* `minikube ip` is the only way to find out minikube's external ip address.
* That's not available from inside the pod,
so the ip has to be given by the user via config,
so that the x-site cluster can form.
* Instead of adding a custom configuration element,
use v1.ServiceSpec externalIP array field for that.
* It reuses existing configuration options,
instead of creating brand new ones.
* Define external site service via .spec.local.expose attribute.
* It avoids reusing the term "service" in the configuration,
given that we already use service to differentiate between
cache and data grid functionality.
* Assume port and nodePort defaults,
when exposed service is of type NodePort.